refactor: use react form hook to update state value
